Move S3 socket env primitive into service
S3 now owns its executable transport environment naming API directly:\n\n    import "github.com/valon-technologies/gestalt/server/services/s3"\n\n    socketEnv := s3.SocketEnv("attachments")\n    tokenEnv := s3.SocketTokenEnv("attachments")\n\nThe config validator now uses the service-owned helper when checking S3 binding env collisions, and the old internal/providerenv wrapper is removed. This keeps the S3 transport contract colocated with the S3 service package while preserving the existing env-name normalization behavior.
